homework2 : - доработал сохранение заметки как с датой так и без нее - добавил функционал удаление текстового поля даты для записи заметки без даты - провел рефакторинг кода в Extension
