+++ Notes/Models/Note.swift
@@ -8,15 +8,11 @@
 import Foundation
 
 struct Note {
-    let title: String?
-    let description: String?
-    let date: Date
+    let title: String
+    let description: String
+    let date: String?
 
     var isEmpty: Bool {
-        guard let title = title, let description = description else {
-            return true
-        }
-
         if title.isEmpty && description.isEmpty {
           return true
         } else {
